Implement and use CardStore CRUD operations for Reviews, Decks, and Cards
Still very unsure on the implementation, will get a working demo then refactor as I see fit

+++ src/activities/flashcards/AnkiConnect.h
@@ -29,7 +29,7 @@ struct CardInfo {
 };
 
 struct Deck {
-  int id;
+  long long id;
   std::string name;
 };
